DEVTASK-20662

+++ app/Http/Controllers/TwillioMessageController.php
@@ -85,8 +85,10 @@ class TwillioMessageController extends Controller{
 		$messageGroupCustomer = MessagingGroupCustomer::where('id', $customerId)->first();
 		if($messageGroupCustomer != null) {
 			$marketing_message = MarketingMessage::where('message_group_id', $messageGroupCustomer['message_group_id'])->first();
-			MarketingMessageCustomer::where(['marketing_message_id'=> $marketing_message->id, 'customer_id'=>$customerId])->delete();
-			$messageGroupCustomer->delete();
+			if($marketing_message != null) {
+				MarketingMessageCustomer::where(['marketing_message_id'=> $marketing_message->id, 'customer_id'=>$customerId])->delete();
+				$messageGroupCustomer->delete();
+			}
 		}
 		return response()->json(['code' => 200, 'message' => 'Customer removed from message group successfully']);
 	}
@@ -97,8 +99,10 @@ class TwillioMessageController extends Controller{
 		if($messageGroup != null) {
 			$customers = MessagingGroupCustomer::where('message_group_id', $request->id)->delete();
 			$marketing_message = MarketingMessage::where('message_group_id', $messageGroupId)->first();
-			MarketingMessageCustomer::where('marketing_message_id', $marketing_message->id)->delete();
-			$marketing_message->delete();
+			if($marketing_message != null) {
+				MarketingMessageCustomer::where('marketing_message_id', $marketing_message->id)->delete();
+				$marketing_message->delete();
+			}
 		}
 		$messageGroup->delete();
 		return response()->json(['code' => 200, 'message' => 'Message group deleted successfully']);
